Refactor: Move tracer class to a new module, tidy up interface with JinjaTemplater
