+++ src/sqlfluff/core/templaters/jinja.py
@@ -3,11 +3,10 @@
 import os.path
 import logging
 import importlib.util
-import re
-import uuid
-from itertools import chain
-from typing import List, Tuple, Optional
+from typing import Callable, Dict, List, Tuple, Optional
 
+from jinja2 import Environment
+from jinja2.environment import Template
 from jinja2.sandbox import SandboxedEnvironment
 from jinja2 import meta, TemplateSyntaxError, TemplateError
 import jinja2.nodes
@@ -20,6 +19,7 @@ from sqlfluff.core.templaters.base import (
     TemplatedFileSlice,
 )
 from sqlfluff.core.templaters.python import PythonTemplater
+from sqlfluff.core.templaters.tracer import JinjaTracer
 
 # Instantiate the templater logger
 templater_logger = logging.getLogger("sqlfluff.templater")
@@ -165,7 +165,7 @@ class JinjaTemplater(PythonTemplater):
             )
 
     @staticmethod
-    def get_jinja_env():
+    def _get_jinja_env():
         """Get a properly configured jinja environment."""
         # We explicitly want to preserve newlines.
         return SandboxedEnvironment(
@@ -174,6 +174,70 @@ class JinjaTemplater(PythonTemplater):
             autoescape=False,
             extensions=["jinja2.ext.do"],
         )
+
+    def get_context(self, fname=None, config=None) -> Dict:
+        """Get the templating context from the config."""
+        # Load the context
+        live_context = super().get_context(fname=fname, config=config)
+        # Apply dbt builtin functions if we're allowed.
+        if config:
+            apply_dbt_builtins = config.get_section(
+                (self.templater_selector, self.name, "apply_dbt_builtins")
+            )
+            if apply_dbt_builtins:
+                # This feels a bit wrong defining these here, they should probably
+                # be configurable somewhere sensible. But for now they're not.
+                # TODO: Come up with a better solution.
+                dbt_builtins = self._generate_dbt_builtins()
+                for name in dbt_builtins:
+                    # Only apply if it hasn't already been set at this stage.
+                    if name not in live_context:
+                        live_context[name] = dbt_builtins[name]
+
+        env = self._get_jinja_env()
+
+        # Load macros from path (if applicable)
+        if config:
+            macros_path = config.get_section(
+                (self.templater_selector, self.name, "load_macros_from_path")
+            )
+            if macros_path:
+                live_context.update(
+                    self._extract_macros_from_path(
+                        macros_path, env=env, ctx=live_context
+                    )
+                )
+
+            # Load config macros, these will take precedence over macros from the path
+            live_context.update(
+                self._extract_macros_from_config(
+                    config=config, env=env, ctx=live_context
+                )
+            )
+
+            live_context.update(self._extract_libraries_from_config(config=config))
+        return live_context
+
+    def template_builder(
+        self, fname=None, config=None
+    ) -> Tuple[Environment, dict, Callable[[str], Template]]:
+        """Builds and returns objects needed to create and run templates."""
+        # Load the context
+        live_context = self.get_context(fname=fname, config=config)
+        env = self._get_jinja_env()
+
+        def make_template(in_str):
+            """Used by JinjaTracer to instantiate templates.
+
+            This function is a closure capturing internal state from process().
+            Note that creating templates involves quite a bit of state known to
+            _this_ function but not to JinjaTracer.
+
+            https://www.programiz.com/python-programming/closure
+            """
+            return env.from_string(in_str, globals=live_context)
+
+        return env, live_context, make_template
 
     def process(
         self, *, in_str: str, fname: str, config=None, formatter=None
@@ -202,42 +266,9 @@ class JinjaTemplater(PythonTemplater):
                 "For the jinja templater, the `process()` method requires a config object."
             )
 
-        # Load the context
-        live_context = self.get_context(fname=fname, config=config)
-        # Apply dbt builtin functions if we're allowed.
-        apply_dbt_builtins = config.get_section(
-            (self.templater_selector, self.name, "apply_dbt_builtins")
+        env, live_context, make_template = self.template_builder(
+            fname=fname, config=config
         )
-        if apply_dbt_builtins:
-            # This feels a bit wrong defining these here, they should probably
-            # be configurable somewhere sensible. But for now they're not.
-            # TODO: Come up with a better solution.
-            dbt_builtins = self._generate_dbt_builtins()
-            for name in dbt_builtins:
-                # Only apply if it hasn't already been set at this stage.
-                if name not in live_context:
-                    live_context[name] = dbt_builtins[name]
-
-        env = self.get_jinja_env()
-
-        # Load macros from path (if applicable)
-        macros_path = config.get_section(
-            (self.templater_selector, self.name, "load_macros_from_path")
-        )
-        if macros_path:
-            live_context.update(
-                self._extract_macros_from_path(macros_path, env=env, ctx=live_context)
-            )
-
-        # Load config macros, these will take precedence over macros from the path
-        live_context.update(
-            self._extract_macros_from_config(config=config, env=env, ctx=live_context)
-        )
-
-        live_context.update(self._extract_libraries_from_config(config=config))
-
-        def make_template(in_str):
-            return env.from_string(in_str, globals=live_context)
 
         # Load the template, passing the global context.
         try:
@@ -245,7 +276,6 @@ class JinjaTemplater(PythonTemplater):
         except TemplateSyntaxError as err:
             # Something in the template didn't parse, return the original
             # and a violation around what happened.
-            (len(line) for line in in_str.split("\n")[: err.lineno])
             return (
                 TemplatedFile(source_str=in_str, fname=fname),
                 [
@@ -317,7 +347,8 @@ class JinjaTemplater(PythonTemplater):
         """Slice the file to determine regions where we can fix."""
         # The TemplateTracer slicing algorithm is more robust, but it requires
         # us to create and render a second template (not raw_str) and is only
-        # enabled if the caller passes a make_template() function.
+        # enabled if the caller passes a make_template() function. (For now,
+        # the dbt templater does not.)
         make_template = kwargs.pop("make_template", None)
         if make_template is None:
             # make_template() was not provided. Use the base class
@@ -327,330 +358,6 @@ class JinjaTemplater(PythonTemplater):
         templater_logger.info("Slicing File Template")
         templater_logger.debug("    Raw String: %r", raw_str)
         templater_logger.debug("    Templated String: %r", templated_str)
-        tracer = TemplateTracer(raw_str, templated_str, make_template)
-        return tracer.process()
-
-
-class TemplateTracer:
-    """Deduces and records execution path of a Jinja template.
-
-    This class should be compatible with other template languages but has only
-    been used with Jinja thus far.
-    """
-
-    re_open_tag = re.compile(r"^\s*({[{%])[\+\-]?\s*")
-    re_close_tag = re.compile(r"\s*[\+\-]?([}%]})\s*$")
-
-    def __init__(self, raw_str, templated_str, make_template):
-        self.raw_str = raw_str
-        self.templated_str = templated_str
-        self.make_template = make_template
-        self.program_counter = 0
-        self.raw_sliced = self._slice_template(self.raw_str)
-        self.sliced_file = []
-        self.source_idx = 0
-
-    def process(self):
-        """Captures the trace."""
-        trace_template_str = "".join(
-            rs.alternate_code if rs.alternate_code is not None else rs.raw
-            for rs in self.raw_sliced
-        )
-        # print(trace_template_str)
-        trace_template = self.make_template(trace_template_str)
-        trace_template_output = trace_template.render()
-        # Split output by section. Each section has two possible formats.
-        for p in trace_template_output.split("\0"):
-            if not p:
-                continue
-            m_id = re.match(r"^([0-9a-f]+)(_(\d+))?", p)
-            if m_id.group(3):
-                # E.g. "2e8577c1d045439ba8d3b9bf47561de3_83". The number after
-                # "_" is the length (in characters) of a corresponding literal
-                # in raw_str.
-                value = [m_id.group(1), int(m_id.group(3)), True]
-            else:
-                # E.g. "adc15d2a41d14ead97411bce3fb55e32 a < 10". The characters
-                # after the UUID are executable code from raw_str.
-                value = [m_id.group(0), p[len(m_id.group(0)) + 1 :], False]
-            alt_id, content_info, literal = value
-            target_slice_idx = self.find_slice_index(alt_id)
-            if literal:
-                self.move_to_slice(target_slice_idx, content_info)
-            else:
-                self.move_to_slice(target_slice_idx, len(str(content_info)))
-        return self.raw_sliced, self.sliced_file, self.templated_str
-
-    def find_slice_index(self, slice_identifier) -> int:
-        """Given a slice identifier (UUID string), return its index."""
-        raw_slices_search_result = [
-            idx
-            for idx, rs in enumerate(self.raw_sliced)
-            if rs.unique_alternate_id == slice_identifier
-        ]
-        if len(raw_slices_search_result) != 1:
-            raise ValueError(
-                f"Internal error. Unable to locate slice for {slice_identifier}."
-            )
-        return raw_slices_search_result[0]
-
-    def move_to_slice(self, target_slice_idx, target_slice_length):
-        """Given a template location, walk execution to that point."""
-        while self.program_counter < len(self.raw_sliced):
-            self.record_trace(
-                target_slice_length if self.program_counter == target_slice_idx else 0
-            )
-            current_raw_slice = self.raw_sliced[self.program_counter]
-            if self.program_counter == target_slice_idx:
-                # Reached the target slice. Go to next location and stop.
-                self.program_counter += 1
-                break
-            elif not current_raw_slice.next_slice_indices:
-                # No choice available. Go to next location.
-                self.program_counter += 1
-            else:
-                # We have choices. Which to choose?
-                candidates = []
-                for next_slice_idx in chain(
-                    current_raw_slice.next_slice_indices, [self.program_counter + 1]
-                ):
-                    if next_slice_idx > target_slice_idx:
-                        # Takes us past the target. No good.
-                        continue
-                    candidates.append(next_slice_idx)
-                # Choose the path that lands us closest to the target.
-                candidates.sort(key=lambda c: abs(target_slice_idx - c))
-                self.program_counter = candidates[0]
-
-    def record_trace(self, target_slice_length):
-        """Add the current location to the trace."""
-        slice_type = self.raw_sliced[self.program_counter].slice_type
-        self.sliced_file.append(
-            TemplatedFileSlice(
-                slice_type,
-                slice(
-                    self.raw_sliced[self.program_counter].source_idx,
-                    self.raw_sliced[self.program_counter + 1].source_idx
-                    if self.program_counter + 1 < len(self.raw_sliced)
-                    else len(self.raw_str),
-                ),
-                slice(self.source_idx, self.source_idx + target_slice_length),
-            )
-        )
-        if slice_type in ("literal", "templated"):
-            self.source_idx += target_slice_length
-
-    @classmethod
-    def _slice_template(cls, in_str: str) -> List[RawFileSlice]:
-        """Slice template in jinja.
-
-        NB: Starts and ends of blocks are not distinguished.
-        """
-        env = JinjaTemplater.get_jinja_env()
-        str_buff = ""
-        idx = 0
-        # We decide the "kind" of element we're dealing with
-        # using it's _closing_ tag rather than it's opening
-        # tag. The types here map back to similar types of
-        # sections in the python slicer.
-        block_types = {
-            "variable_end": "templated",
-            "block_end": "block",
-            "comment_end": "comment",
-            # Raw tags should behave like blocks. Note that
-            # raw_end and raw_begin are whole tags rather
-            # than blocks and comments where we get partial
-            # tags.
-            "raw_end": "block",
-            "raw_begin": "block",
-        }
-
-        # https://jinja.palletsprojects.com/en/2.11.x/api/#jinja2.Environment.lex
-        stack = []
-        result = []
-        set_idx = None
-        unique_alternate_id: Optional[str]
-        alternate_code: Optional[str]
-        for _, elem_type, raw in env.lex(in_str):
-            # Replace literal text with a unique ID, except for "set"
-            # statements, which don't emit output and thus don't need this
-            # treatment.
-            if elem_type == "data":
-                if set_idx is None:
-                    unique_alternate_id = uuid.uuid4().hex
-                    alternate_code = f"{unique_alternate_id}_{len(raw)}\0"
-                else:
-                    unique_alternate_id = None
-                    alternate_code = None
-                result.append(
-                    RawFileSlice(
-                        raw,
-                        "literal",
-                        idx,
-                        unique_alternate_id=unique_alternate_id,
-                        alternate_code=alternate_code,
-                    )
-                )
-                idx += len(raw)
-                continue
-            str_buff += raw
-
-            if elem_type.endswith("_begin"):
-                # When a "begin" tag (whether block, comment, or data) uses
-                # whitespace stripping
-                # (https://jinja.palletsprojects.com/en/3.0.x/templates/#whitespace-control),
-                # the Jinja lex() function handles this by discarding adjacent
-                # whitespace from in_str. For more insight, see the tokeniter()
-                # function in this file:
-                # https://github.com/pallets/jinja/blob/main/src/jinja2/lexer.py
-                # We want to detect and correct for this in order to:
-                # - Correctly update "idx" (if this is wrong, that's a
-                #   potential DISASTER because lint fixes use this info to
-                #   update the source file, and incorrect values often result in
-                #   CORRUPTING the user's file so it's no longer valid SQL. :-O
-                # - Guarantee that the slices we return fully "cover" the
-                #   contents of in_str.
-                #
-                # We detect skipped characters by looking ahead in in_str for
-                # the token just returned from lex(). The token text will either
-                # be at the current 'idx' position (if whitespace stripping did
-                # not occur) OR it'll be farther along in in_str, but we're
-                # GUARANTEED that lex() only skips over WHITESPACE; nothing else.
-
-                # Find the token returned. Did lex() skip over any characters?
-                num_chars_skipped = in_str.index(raw, idx) - idx
-                if num_chars_skipped:
-                    # Yes. It skipped over some characters. Compute a string
-                    # containing the skipped characters.
-                    skipped_str = in_str[idx : idx + num_chars_skipped]
-
-                    # Sanity check: Verify that Jinja only skips over
-                    # WHITESPACE, never anything else.
-                    if not skipped_str.isspace():  # pragma: no cover
-                        templater_logger.warning(
-                            "Jinja lex() skipped non-whitespace: %s", skipped_str
-                        )
-                    # Treat the skipped whitespace as a literal.
-                    result.append(
-                        RawFileSlice(skipped_str, "literal", idx, alternate_code="")
-                    )
-                    idx += num_chars_skipped
-
-            # raw_end and raw_begin behave a little differently in
-            # that the whole tag shows up in one go rather than getting
-            # parts of the tag at a time.
-            unique_alternate_id = None
-            alternate_code = None
-            if elem_type.endswith("_end") or elem_type == "raw_begin":
-                block_type = block_types[elem_type]
-                block_subtype = None
-                # Handle starts and ends of blocks
-                if block_type in ("block", "templated"):
-                    # Trim off the brackets and then the whitespace
-                    m_open = cls.re_open_tag.search(str_buff)
-                    m_close = cls.re_close_tag.search(str_buff)
-                    trimmed_content = ""
-                    if m_open and m_close:
-                        trimmed_content = str_buff[
-                            len(m_open.group(0)) : -len(m_close.group(0))
-                        ]
-                    if block_type == "block":
-                        if trimmed_content.startswith("end"):
-                            block_type = "block_end"
-                        elif trimmed_content.startswith("el"):
-                            # else, elif
-                            block_type = "block_mid"
-                        else:
-                            block_type = "block_start"
-                            if trimmed_content.split()[0] == "for":
-                                block_subtype = "loop"
-                    else:
-                        # For "templated", evaluate the content in case of side
-                        # effects, but return a UUID.
-                        if trimmed_content:
-                            assert m_open and m_close
-                            unique_id = uuid.uuid4().hex
-                            unique_alternate_id = unique_id
-                            alternate_code = f"{unique_alternate_id} {m_open.group(1)} {trimmed_content} {m_close.group(1)}\0"
-                if block_type == "block_start" and trimmed_content.split()[0] == "set":
-                    # Jinja supports two forms of {% set %}:
-                    # - {% set variable = value %}
-                    # - {% set variable %}value{% endset %}
-                    # https://jinja.palletsprojects.com/en/2.10.x/templates/#block-assignments
-                    # When the second format is used, leave the value "as is".
-                    # It won't be rendered directly to the template output
-                    # anyway, so substituting our special UUID values would just
-                    # confuse things.
-                    trimmed_content_parts = trimmed_content.split(maxsplit=2)
-                    if len(trimmed_content_parts) <= 2 or not trimmed_content_parts[
-                        2
-                    ].startswith("="):
-                        set_idx = len(result)
-                elif block_type == "block_end" and set_idx is not None:
-                    set_idx = None
-                m = re.search(r"\s+$", raw, re.MULTILINE | re.DOTALL)
-                if raw.startswith("-") and m:
-                    # Right whitespace was stripped. Split off the trailing
-                    # whitespace into a separate slice. The desired behavior is
-                    # to behave similarly as the left stripping case above.
-                    # Note that the stakes are a bit different, because lex()
-                    # hasn't *omitted* any characters from the strings it
-                    # returns, it has simply grouped them differently than we
-                    # want.
-                    trailing_chars = len(m.group(0))
-                    result.append(
-                        RawFileSlice(
-                            str_buff[:-trailing_chars],
-                            block_type,
-                            idx,
-                            block_subtype,
-                            unique_alternate_id=unique_alternate_id,
-                            alternate_code=alternate_code,
-                        )
-                    )
-                    block_idx = len(result) - 1
-                    idx += len(str_buff) - trailing_chars
-                    result.append(
-                        RawFileSlice(
-                            str_buff[-trailing_chars:],
-                            "literal",
-                            idx,
-                            alternate_code="",
-                        )
-                    )
-                    idx += trailing_chars
-                else:
-                    result.append(
-                        RawFileSlice(
-                            str_buff,
-                            block_type,
-                            idx,
-                            block_subtype,
-                            unique_alternate_id=unique_alternate_id,
-                            alternate_code=alternate_code,
-                        )
-                    )
-                    block_idx = len(result) - 1
-                    idx += len(str_buff)
-                if block_type == "block_start" and trimmed_content.split()[0] in (
-                    "for",
-                    "if",
-                ):
-                    stack.append(block_idx)
-                elif block_type == "block_mid":
-                    # Record potential forward jump over this block.
-                    result[stack[-1]].next_slice_indices.append(block_idx)
-                    stack.pop()
-                    stack.append(block_idx)
-                elif block_type == "block_end" and trimmed_content.split()[0] in (
-                    "endfor",
-                    "endif",
-                ):
-                    # Record potential forward jump over this block.
-                    result[stack[-1]].next_slice_indices.append(block_idx)
-                    if result[stack[-1]].slice_subtype == "loop":
-                        # Record potential backward jump to the loop beginning.
-                        result[block_idx].next_slice_indices.append(stack[-1] + 1)
-                    stack.pop()
-                str_buff = ""
-        return result
+        tracer = JinjaTracer(raw_str, cls._get_jinja_env(), make_template)
+        trace = tracer.trace()
+        return trace.raw_sliced, trace.sliced_file, trace.templated_str
